Trukket ut og refaktorert komponentene knyttet til multiselect.
