+++ src/frontend/typer/vedtak.ts
@@ -1,4 +1,8 @@
-import { VilkårType } from './vilkår';
+import navFarger from 'nav-frontend-core';
+
+import { Ressurs, RessursStatus } from '@navikt/familie-typer';
+
+import { Vilkårsbegrunnelser, VilkårType } from './vilkår';
 
 export interface IVedtakForBehandling {
     aktiv: boolean;
@@ -61,4 +65,46 @@ export const vedtakBegrunnelseTyper: Record<VedtakBegrunnelseType, string> = {
     INNVILGELSE: 'Innvilgelse',
     REDUKSJON: 'Reduksjon',
     OPPHØR: 'Opphør',
+};
+
+export const finnVedtakBegrunnelseType = (
+    vilkårBegrunnelser: Ressurs<Vilkårsbegrunnelser>,
+    vedtakBegrunnelse: VedtakBegrunnelse
+): VedtakBegrunnelseType | undefined => {
+    return vilkårBegrunnelser.status === RessursStatus.SUKSESS
+        ? (Object.keys(vilkårBegrunnelser.data).find(vedtakBegrunnelseType => {
+              return (
+                  vilkårBegrunnelser.data[vedtakBegrunnelseType as VedtakBegrunnelseType].find(
+                      (vedtakBegrunnelseTilknyttetVilkår: IRestVedtakBegrunnelseTilknyttetVilkår) =>
+                          vedtakBegrunnelseTilknyttetVilkår.id === vedtakBegrunnelse
+                  ) !== undefined
+              );
+          }) as VedtakBegrunnelseType)
+        : undefined;
+};
+
+export const hentBakgrunnsfarge = (vedtakBegrunnelseType?: VedtakBegrunnelseType) => {
+    switch (vedtakBegrunnelseType) {
+        case VedtakBegrunnelseType.INNVILGELSE:
+            return navFarger.navGronnLighten80;
+        case VedtakBegrunnelseType.REDUKSJON:
+            return navFarger.navOransjeLighten80;
+        case VedtakBegrunnelseType.OPPHØR:
+            return navFarger.navLysGra;
+        default:
+            return navFarger.navBlaLighten80;
+    }
+};
+
+export const hentBorderfarge = (vedtakBegrunnelseType?: VedtakBegrunnelseType) => {
+    switch (vedtakBegrunnelseType) {
+        case VedtakBegrunnelseType.INNVILGELSE:
+            return navFarger.navGronn;
+        case VedtakBegrunnelseType.REDUKSJON:
+            return navFarger.navOransjeDarken20;
+        case VedtakBegrunnelseType.OPPHØR:
+            return navFarger.navGra60;
+        default:
+            return navFarger.navBlaLighten80;
+    }
 };
